yoshino: clean ramdump partition for userdebug and eng variants
This is to make sure ramdumper won't be invoked for the userdebug and
engineering variants. Service is started from init.$(TARGET_DEVICE).rc
right after the mount_all, check the $OUT/root/. There must be no
traces of this tool in the user variant.

NOTE: current version is for the system/vendor/ and has to be revised
      once vendor image is enabled

Signed-off-by: Oleksiy Avramchenko <oleksiy.avramchenko@sony.com>