platform: Do not share specific audio policy bits between CAF and AOSP
Creating a shared policy base between AOSP and CAF is almost impossible
as we try to override and update more and more nodes such as extending
existing mixPorts. While moving such nodes into a separate file for AOSP
and CAF (to be more explicit: merge caf_common and caf_routes together,
then create a similar aosp_common file) is definitely possible, the
resulting set of shared nodes is relatively small for the extra
management overhead it costs us to compare and split these out.

Instead, duplicate the entire policy: freeze AOSP in place and create
a copy for CAF. Note that this temporarily does away with the PARTIAL
CAF policy files in favour of importing them all at once in a clean
followup commit. That policy was anyway rejected by APM, see early
pushes to https://github.com/sonyxperiadev/device-sony-kumano/pull/38.
With the changes in a separate commit the total diff from AOSP to CAF is
much easier to review.

This makes it easier to manage current-and-future top-level
audio_policy_configuration files too. As a concrete example
audio_policy_configuration_bluetooth_legacy_hal.xml wasn't updated to
xi:include the new files, and the same would have happened to
(future) audio_policy_configuration_a2dp_offload_disabled.xml.

In the end the AOSP policy persists only for future reference and as
fallback, but is likely to be removed regardless; the CAF audio HAL is
by far superior. Testing and maintaining this policy is overhead.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>