audio_io_policy: Replace with YOSHINO-2.2.0-190725-0908
The file added in #159 came straight from Kumano; this may unfortunately
be invalid for Yoshino; hence import the right one from stock.
Note that the file was originally named audio_output_policy, which is
an older filename [1] (a name change made to accomodate for a new block
inputs{}, used to configure input ports).

Signed-off-by: Marijn Suijten <marijns95@gmail.com>

[1]: https://github.com/sonyxperiadev/vendor-qcom-opensource-audio-hal-primary-hal/blob/4943b04bfab800925d62c318763f5892221acd82/hal/audio_extn/utils.c#L539-L543