caf_audio_policy: Replace BT SCO All with devicePorts for every mode
The All node was imported from stock, replacing three more specific
nodes. Unfortunately this leads to a segfault in APM when connecting to
bluetooth headphones:

https://github.com/sonyxperiadev/device-sony-yoshino/pull/161#issuecomment-650719993

Hence go back to the previous routing layout, that conveniently matches
newer platforms on stock. The devicePorts themselves have not been
removed on stock anyway, only the routes.

(Note that, as tested, this is not caused by the existing but
unreferenced devicePorts for BT SCO, BT SCO Headset and BT SCO Car Kit,
but by the mere existence of a sink route with BT SCO All)

Signed-off-by: Marijn Suijten <marijns95@gmail.com>